Add on_demand flag to exes

+++ setup2/process.py
@@ -48,6 +48,7 @@ async def fetch_file(url: str, version: Optional[str]) -> str:
     _, file = path.split(full_url)
 
     filename = f'{conf.sources_dir}/{file}'
-    await async_proc(f'curl -L {full_url} -o {filename}')
+    if not path.exists(filename):
+        await async_proc(f'curl -L {full_url} -o {filename}')
 
     return filename
